backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==40.0.2
daphne==4.0.0
dataclasses-json==0.5.7
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
frozenlist==1.3.3
greenlet==2.0.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
langchain==0.0.166
lxml==4.9.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.6
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.1
pdfkit==1.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
pyOpenSSL==23.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==4.5.4
requests==2.29.0
service-identity==21.1.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.13
sqlparse==0.4.4
tenacity==8.2.2
tqdm==4.65.0
Twisted==22.10.0
txaio==23.1.1
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
yarl==1.9.2
zope.interface==6.0
